long double changed for read  float variable

+++ EPOS_Linux_Library/examples/HelloEposCmd/profile_velocity_mode_CAN.cpp
@@ -60,7 +60,7 @@ const string g_programName = "HelloEposCmd";
 //Profile Velocity Default Inputs
 long targetvelocity_1 = 0; //rpm
 long targetvelocity_2 = 0; //rpm
-int TargetTorqueNode2 = 0; //1000x * Motor Rated Torque
+long double TargetTorqueNode2 = 0; //1000x * Motor Rated Torque
 long double simtime = 0; //sec
 /* vector<double> p_CurrentIs_saved;
 vector<double> p_Time_saved; */
